+++ src/ObjectCollection/DateTimeImmutableNullableCollection.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Steevanb\PhpCollection\ObjectCollection;
+
+/** @extends AbstractObjectNullableCollection<\DateTimeImmutable|null> */
+class DateTimeImmutableNullableCollection extends AbstractObjectNullableCollection
+{
+    public static function getValueFqcn(): string
+    {
+        return \DateTimeImmutable::class;
+    }
+}