+++ src/ObjectCollection/ReflectionAttributeNullableCollection.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Steevanb\PhpCollection\ObjectCollection;
+
+/** @extends AbstractObjectNullableCollection<\ReflectionAttribute<object>|null> */
+class ReflectionAttributeNullableCollection extends AbstractObjectNullableCollection
+{
+    public static function getValueFqcn(): string
+    {
+        return \ReflectionAttribute::class;
+    }
+}