+++ src/ObjectCollection/ReflectionClassNullableCollection.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Steevanb\PhpCollection\ObjectCollection;
+
+/** @extends AbstractObjectNullableCollection<\ReflectionClass<object>|null> */
+class ReflectionClassNullableCollection extends AbstractObjectNullableCollection
+{
+    public static function getValueFqcn(): string
+    {
+        return \ReflectionClass::class;
+    }
+}