+++ src/ObjectCollection/ReflectionIntersectionTypeNullableCollection.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Steevanb\PhpCollection\ObjectCollection;
+
+/** @extends AbstractObjectNullableCollection<\ReflectionIntersectionType|null> */
+class ReflectionIntersectionTypeNullableCollection extends AbstractObjectNullableCollection
+{
+    public static function getValueFqcn(): string
+    {
+        return \ReflectionIntersectionType::class;
+    }
+}