+++ src/ObjectCollection/ReflectionMethodNullableCollection.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Steevanb\PhpCollection\ObjectCollection;
+
+/** @extends AbstractObjectNullableCollection<\ReflectionMethod|null> */
+class ReflectionMethodNullableCollection extends AbstractObjectNullableCollection
+{
+    public static function getValueFqcn(): string
+    {
+        return \ReflectionMethod::class;
+    }
+}