+++ src/ObjectCollection/SplFileObjectCollection.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Steevanb\PhpCollection\ObjectCollection;
+
+/** @extends AbstractObjectCollection<\SplFileObject> */
+class SplFileObjectCollection extends AbstractObjectCollection
+{
+    public static function getValueFqcn(): string
+    {
+        return \SplFileObject::class;
+    }
+}